% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.geneset.name.R
\name{geneset.names}
\alias{geneset.names}
\title{Print all geneset names in geneset list}
\usage{
geneset.names(geneset_list = NULL)
}
\arguments{
\item{geneset_list}{User provided genesets list loaded by load.geneset(). Phanotm package provids four geneset lists from different resources: kegg, reactome, emory geneset and baylor modules. These genesets can be obtained with data(), e.g. data(kegg.geneset)}
}
\description{
This function allows user to print all the geneset names in the loaded geneset list such that user can find the specific query_geneset name for run.phantom funciton.
}
\examples{

## store all the geneset names in one vector
 \dontrun{
g.names = geneset.names(reactome.geneset)
}
}
\keyword{geneset}
\keyword{names}
\keyword{print}
