% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.data.R
\name{load.data}
\alias{load.data}
\title{Load data function}
\usage{
load.data(filename)
}
\arguments{
\item{filename}{the time-course gene differential expression data file path and name on user's local machine}
}
\description{
This function allows user to load in the time-course gene differential expression data in a format that can be used by
run.phantom and run.phantom.batch functions.
}
\keyword{data}
\keyword{load}
