
\name{reactome.geneset}
\docType{data}
\alias{reactome.geneset}
\title{reactome genesets}
\description{
  This data set gives the a geneset list of reactome gene sets
}
\usage{reactome.geneset
}
\format{A large list containing 674 gene sets.}
\source{reactome genesets from MSigDB}
\keyword{datasets}
