
\name{time.course.data}
\docType{data}
\alias{time.course.data}
\title{A sample time.course.data with 15 time points}
\description{
  This data set gives a sample of time-course gene differential expression data, designated by t statistics of symtomatic samples versus asymtomatic samples gene expression across 15 time points.
}
\usage{time.course.data
}
\format{A large matrix containing the t statistics of 11745 genes across 15 time points}
\source{publications}
\keyword{datasets}
